Functionality for percentage calculation of successful reviews

+++ src/main/java/org/xenon/echo/repositories/ReviewRepository.java
@@ -29,5 +29,5 @@ public interface ReviewRepository extends JpaRepository<Review,Long> {
     @Query("""
       SELECT COUNT(r) FROM Review  r WHERE r.memoryItem.user.id =:userId AND r.rating <> org.xenon.echo.enums.ReviewRating.AGAIN AND r.rating <> org.xenon.echo.enums.ReviewRating.HARD
     """)
-    long countSuccessfulReviews(UUID userId);
+    double countSuccessfulReviews(UUID userId);
 }
